+++ lib/deimos/consumer.rb
@@ -1,6 +1,7 @@
 # frozen_string_literal: true
 
 require 'deimos/avro_data_decoder'
+require 'deimos/base_consumer'
 require 'deimos/shared_config'
 require 'phobos/handler'
 require 'active_support/all'
@@ -11,23 +12,8 @@ require 'ddtrace'
 # for every incoming message. This class should be lightweight.
 module Deimos
   # Parent consumer class.
-  class Consumer
+  class Consumer < BaseConsumer
     include Phobos::Handler
-    include SharedConfig
-
-    class << self
-      # @return [AvroDataEncoder]
-      def decoder
-        @decoder ||= AvroDataDecoder.new(schema: config[:schema],
-                                         namespace: config[:namespace])
-      end
-
-      # @return [AvroDataEncoder]
-      def key_decoder
-        @key_decoder ||= AvroDataDecoder.new(schema: config[:key_schema],
-                                             namespace: config[:namespace])
-      end
-    end
 
     # :nodoc:
     def around_consume(payload, metadata)
@@ -41,31 +27,8 @@ module Deimos
     # :nodoc:
     def before_consume(payload, metadata)
       _with_error_span(payload, metadata) do
-        if self.class.config[:key_schema] || self.class.config[:key_field]
-          metadata[:key] = decode_key(metadata[:key])
-        end
+        metadata[:key] = decode_key(metadata[:key])
         self.class.decoder.decode(payload) if payload.present?
-      end
-    end
-
-    # Helper method to decode an Avro-encoded key.
-    # @param key [String]
-    # @return [Object] the decoded key.
-    def decode_key(key)
-      return nil if key.nil?
-
-      config = self.class.config
-      if config[:encode_key] && config[:key_field].nil? &&
-         config[:key_schema].nil?
-        raise 'No key config given - if you are not decoding keys, please use `key_config plain: true`'
-      end
-
-      if config[:key_field]
-        self.class.decoder.decode_key(key, config[:key_field])
-      elsif config[:key_schema]
-        self.class.key_decoder.decode(key, schema: config[:key_schema])
-      else # no encoding
-        key
       end
     end
 
@@ -78,20 +41,6 @@ module Deimos
 
   private
 
-    # @param payload [Hash|String]
-    # @param metadata [Hash]
-    def _with_error_span(payload, metadata)
-      @span = Deimos.config.tracer&.start(
-        'deimos-consumer',
-        resource: self.class.name.gsub('::', '-')
-      )
-      yield
-    rescue StandardError => e
-      _handle_error(e, payload, metadata)
-    ensure
-      Deimos.config.tracer&.finish(@span)
-    end
-
     def _received_message(payload, metadata)
       Deimos.config.logger.info(
         message: 'Got Kafka event',
@@ -102,13 +51,13 @@ module Deimos
                                          status:received
                                          topic:#{metadata[:topic]}
                                        ))
+      _report_time_delayed(payload, metadata)
     end
 
     # @param exception [Throwable]
     # @param payload [Hash]
     # @param metadata [Hash]
     def _handle_error(exception, payload, metadata)
-      Deimos.config.tracer&.set_error(@span, exception)
       Deimos.config.metrics&.increment(
         'handler',
         tags: %W(
@@ -124,7 +73,7 @@ module Deimos
         error_message: exception.message,
         error: exception.backtrace
       )
-      raise if Deimos.config.reraise_consumer_errors
+      super
     end
 
     # @param time_taken [Float]
@@ -145,20 +94,6 @@ module Deimos
         time_elapsed: time_taken,
         metadata: metadata
       )
-      return if payload.nil? || payload['timestamp'].blank?
-
-      begin
-        time_delayed = Time.now.in_time_zone - payload['timestamp'].to_datetime
-      rescue ArgumentError
-        Deimos.config.logger.info(
-          message: "Error parsing timestamp! #{payload['timestamp']}"
-        )
-        return
-      end
-      Deimos.config.metrics&.histogram('handler', time_delayed, tags: %W(
-                                         time:time_delayed
-                                         topic:#{metadata[:topic]}
-                                       ))
     end
   end
 end