+++ spec/handlers/my_batch_consumer.rb
@@ -1,0 +1,5 @@
+# frozen_string_literal: true
+
+module ConsumerTest
+  class MyBatchConsumer < Deimos::BatchConsumer; end
+end