+++ spec/handlers/my_consumer.rb
@@ -1,0 +1,5 @@
+# frozen_string_literal: true
+
+module ConsumerTest
+  class MyConsumer < Deimos::Consumer; end
+end